+++ changelog.md
@@ -9,6 +9,7 @@
 
 - Fixed issues with logic to restart the vlc player sync loops in the mobile app after returning from background
 - Added sticky back to top to vlc player
+- Rebuilt GRoot menu 
 
 #### v8.11